Split workspace sidebar parity gates
